docs(hello,world,hello-world,archetype): show installation and usage in component READMEs
Each README has a link back to the project README, instructions to install with npm, and sample
usage in jsx. The installation instructions need to have the peer dependencies so that people won't
be hit with errors if they just try to install the composite component. The archetype README is a
sample README for anyone who wants to create a new component. With the prompt in here, hopefully
people will be able to see how they should write READMEs for any new components they write.