build(webpack): enable scope hoisting
Scope Hoisting is the flagship feature of webpack 3. One of webpack’s trade-offs when bundling was
that each module in your bundle would be wrapped in individual function closures. These wrapper
functions made it slower for your JavaScript to execute in the browser. In comparison, tools like
Closure Compiler and RollupJS ‘hoist’ or concatenate the scope of all your modules into one closure
and allow for your code to have a faster execution time in the browser.