build(webpack): display optimization bailouts
Scope Hoisting is specifically a feature made possible by ECMAScript Module syntax. Because of this
webpack may fallback to normal bundling based on what kind of modules we are using, and other
conditions. To stay informed on what triggers these fallbacks, I've added this flag which will tell
us what caused the fallback