fix(site): single component demo routing
running the site for a single component demo was broken after switching the routing to use a hash
lookup because a single demo object wasn't also keyed by it's slug. now by adding the slug back into
the demo object, if the demo object passed into the app has a slug, it's a single demo. otherwise,
it's assumed to be an object where the key is the slug and the value is a demo object.