chore(site): update the way the site is deployed locally.
Deploying to netlify used to be a little flaky from the command line since you had to remember to
authorize the build in the netlify UI at build time, otherwise you get an auth failure, the build
would fail, and the site would not deploy via the netlify CI. Running the netlifyctl tool from the
cli required installing from homebrew, and was needlessly complex. This PR removes the deployment steps from the repo completely, and now we will just publish the site from the netlify UI by choosing a Deploy Preview manually.

connects #320