fix(dependencies): Update minimum React/ReactDOM peer dependency version
It is likely that Mineral UI works in a lower version of React as well, but we do not currently have the infrastructure to test/support it.  Anecdotally, all tests passed and package was fully functional in React 15.4.2.  We may update this version to be more accurate in the future as we develop more automated testing solutions in this area.