build(test): Add happo.io test suite
Happo is a visual regression testing service. It takes screenshots and compares
with previous versions of a component.

The happo suite for mineral-ui consists of the same examples used on the public
website, with the exception of <Popover> and <Dropdown>. These components have
dependencies that make rendering in the happo.io environment challenging. More
on that here: https://github.com/enduire/happo.io#can-happo-render-any-component

To start with, I've added these browser targets to the mix:

- Chrome on an 800x600 window, large enough to test most scenarios,
small enough to not waste pixels on the screenshots.
- Chrome on a 320x640 window, roughly equal to an old smartphone.
- Internet Explorer (version 11) on an 800x600 window.

The last target is likely the most important one, at least it's the
target I see most bugs being caught.

This commit only installs happo and sets up the config necessary in order to
make it run. An upcoming commit will add the necessary CI config and code so
that happo starts running automatically for every PR pushed to mineral-ui.