build(test): Add happo CI integration
This adds a happo.io CI integration with a script that's run on all PRs.
It will
a) check if the baseline commit has an existing report
b) if not, run happo to generate a baseline report
c) run happo to generate a report for the current commit (HEAD of PR)
d) compare the two reports and post a comment to the PR with the results

The comment will either say "No differences found" or give you a summary
of the diffs with a link to a page where you can review the diffs.

A few environment variables (secrets) are needed along the way:
- HAPPO_KEY and HAPPO_SECRET which are used to identify and authenticate
the account on the happo.io side
- GHE_TOKEN which is used when posting the PR comment. For now, this is
going to be a token associated with my @trotzig user. To begin with,
this is probably good, since I get pinged if people reply to it, allows
me to keep an eye on things. In the future, I might ask some admin to
provide a new token.