build(test): Avoid happo failure on first run
Happo is designed to run twice. Once on a baseline commit, then again on
the HEAD commit. In the PR that I'm working on now to add happo support,
the baseline commit won't have happo installed. So trying to `npm run
happo` will always fail.

As a workaround, I'm adding a quick check to see if the happo.io package
is installed before generating the report.

At some point in the near future, I'm going to move this happo-ci.sh
script to the main happo.io repo so that others can benefit from it. But
for now I'm keeping things local to mineral-ui.