build(test): Avoid posting happo comment when no diffs
I've noticed that the comment posted from the happo-ci.sh script can be
noisy, especially for PRs that are pushed to/rebased many times. By
silencing out the comment so that it's only posted when there are one or
more diffs/added/deleted we can avoid noise in many cases.