WIP: I'm pretty sure I got the prop composition right here. I'm less certain about the event handler composition, but I think it's very close if not correct.
I left a bunch of console.logs in there for you to check out.

It turns out that the upstream/downstream didn't make sense because only one
would be defined at a time. So I went for this approach instead. Select is
always a 'listbox' now.

I thought I would have to allow one of the cases to override the event handlers,
but as far as I can tell, everything works as expected with the events as they
are now. If you can construct an experiment to prove/disprove that, that would
be awesome.