test(happo): Update to happo.io 1.2.0
This version has a few improvements compared to 1.1.0. Those that
mineral-ui will most likely benefit from are:

- Local images (e.g. url(/foo.png)) found in CSS will now be part of the
screenshot (technically, the happo client inlines them as base64).

- No need for a customizeWebpackConfig function in config. Happo now has
a babel loader as the default config.

- Better error messages. Instead of seeing an error in a call to `eval`,
you now see the component and variant where the error happened, plus a
location in a js file (this is still the tmp file which isn't ideal, but
better than before at least).

- A polyfill for requestAnimationFrame will get rid of warnings from
React during rendering.

Full list of changes is available here:
https://github.com/enduire/happo.io/releases/tag/v1.2.0