test(happo): Update to happo.io@1.5.0
Recently, I've been working on making Happo runs work for PRs coming
from forks. Version 1.5.0 of the happo.io client has support for
temporary authentication tied to a specific PR on github. This is
helpful when running happo in CI (Travis/CircleCI/etc) because in
general, secret environment variables aren't sent to the job running for
the fork (for obvious reasons: anyone would be able to extract secrets).